block: Add ZEN I/O scheduler
Zen is a highly efficient I/O scheduler for mobile devices, which is based on
Noop, Deadline, SIO and V(R).

It is a FCFS-based scheduler without sorting. It uses deadlines for fairness
and treats synchronous requests with priority over asynchronous ones.

Use it by default, as the tests showed that this is preety much the best I/O
scheduler for eMMC.

Signed-off-by: Alex Saiko <solcmdr@gmail.com>
Signed-off-by: Brandon Berhent <bbedward@gmail.com>